backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend server for Golden Abundance"
authors = [{ name = "Jet Kan", email = "jetkan.yk@gmail.com" }]
requires-python = ">=3.14"
dependencies = [
  "alembic>=1.18.4",
  # Avoid `fastapi[standard]` — it pulls fastapi-cli, jinja2, python-multipart,
  # and pyyaml that aren't needed at runtime (the dev group below carries
  # fastapi-cli for `just dev`). email-validator is kept because
  # `pydantic.EmailStr` needs it at import time (User contract).
  "email-validator>=2.0.0",
  "fastapi>=0.136.0",
  "greenlet>=3.4.0",
  "psycopg[binary]>=3.3.3",
  "pydantic-settings>=2.13.1",
  "pyjwt[crypto]>=2.12.1",
  "sentry-sdk[fastapi]>=2.17.0",
  "slowapi>=0.1.9",
  "sqlmodel>=0.0.38",
  "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
  "fastapi-cli>=0.0.5",
  "httpx>=0.28.1",
  "pytest>=9.0.3",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.1.0",
  "ruff>=0.15.11",
  "testcontainers>=4.14.2",
  "ty>=0.0.31",
]

[tool.fastapi]
entrypoint = "backend.server:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
# ``--cov=backend`` is explicit: the bare ``--cov`` form relies on
# ``tool.coverage.run.source`` and reads magically.
addopts = ["-v", "--cov=backend"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = [
  "src/backend/contract/validate_examples.py",
  "src/backend/scripts/*",
  "src/backend/seed_reset.py",
]

[tool.coverage.report]
fail_under = 90
show_missing = true

# SQLAlchemy's InstrumentedAttribute descriptors (`Row.col == value`,
# `.is_()`, `.in_()`, `.asc()`, `.desc()`, `.ilike()`) and SQLModel's
# dynamically-synthesized `__init__` don't survive ty's type resolution,
# producing false positives in query code. Scope the override to the
# layers that actually touch SQLAlchemy/SQLModel (DB, services, routers,
# seed scripts, and tests) so auth/, config, contract/, and scripts/
# still get full type checking.
[[tool.ty.overrides]]
include = [
  "src/backend/db/**/*.py",
  "src/backend/services/**/*.py",
  "src/backend/routers/**/*.py",
  "src/backend/seed.py",
  "tests/**/*.py",
]

[tool.ty.overrides.rules]
invalid-argument-type = "ignore"
missing-argument = "ignore"
no-matching-overload = "ignore"
unresolved-attribute = "ignore"
unsupported-operator = "ignore"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
  "E", "F", "W", "I", "UP", "B", "SIM", "RUF", "N", "C4", "PTH",
  "ASYNC", "DTZ", "PT", "LOG", "G", "RET", "PIE", "TID",
]
# B008: FastAPI's `= Depends(...)` / `= Query(...)` defaults are idiomatic,
#       not the anti-pattern this rule targets. Project-wide because Depends
#       can appear in any module, not just routers.
ignore = ["B008"]

# Scope ambiguous-unicode ignores to the specific files that legitimately
# need non-ASCII characters (zh-Hant seed content, en-dash in docstrings)
# so the rest of the codebase still catches typo/lookalike-attack cases.
[tool.ruff.lint.per-file-ignores]
"src/backend/seed.py" = ["RUF001", "RUF002", "RUF003"]
"src/backend/services/display_id.py" = ["RUF002"]

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"
